Docker Development Environment Operational.
By cloning the repository and downloading docker, now you
can from the command line run in the repo directory
'docker-compose up' and all environments will be configured
and up and running in a couple of minutes. This includes
-node
-flask
-postgres

You will have to move an uncompressed copy of the pg_dump
file manually into the ./init.d/db/SQL folder before running
the above command for the database to get auto-populated.