Remove Docker configuration
I don't use Docker for development on this project, and no one else is
active, so delete the Docker configuration so that it doesn't get broken
and out-of-sync.

Thank you again, @botwhytho, for your contributions here! I'm glad you
were able to find a way you were happy to work and learn. Although the
files are deleted, they will remain in our history forever.