Allow users to guess a person's name in several variants
Allowed options are: Any of the first names, middle names (if they are
in parens), a combination in the correct order of these, and optionally,
users can also add the last name. Fixes #39.

Generation of variants is done harnessing a power_set function. Tests
check that this function, as well as the name guessing, works as
intended.

Co-authored-by: Jaryn Colbert <jaryn.colbert@gmail.com>