Upgrade Flask and Jinja to address CVE
The version of Jinja we were using, 2.10.0, contains
CVE-2019-10906[1][2] in sandbox mode, which is functionality I don't
believe we are using. Upgrade Jinja to 2.10.1 to address the CVE, and
since Jinja is a dependency of Flask, also upgrade Flask
to version 1.1.1. There don't appear to be any code changes needed for
the new version of Flask[3].

[1] https://nvd.nist.gov/vuln/detail/CVE-2019-10906
[2] https://palletsprojects.com/blog/jinja-2-10-1-released/
[3] https://flask.palletsprojects.com/en/1.1.x/changelog/#version-1-1-1